+++ R/oauth-flow-auth-code.R
@@ -32,7 +32,8 @@
 #' @param cache_disk Should the access token be cached on disk? This reduces
 #'   the number of times that you need to re-authenticate at the cost of
 #'   storing access credentials on disk. Cached tokens are encrypted,
-#'   automatically deleted 30 days after creation, and stored in [cache_path()].
+#'   automatically deleted 30 days after creation, and stored in
+#'   [oauth_cache_path()].
 #' @param cache_key If you want to cache multiple tokens per app, use this
 #'   key to disambiguate them.
 #' @returns A modified HTTP [request].