+++ R/oauth.R
@@ -97,7 +97,7 @@ cache_mem <- function(client, key) {
   )
 }
 cache_disk <- function(client, key) {
-  app_path <- file.path(rappdirs::user_cache_dir("httr2"), client$name)
+  app_path <- file.path(oauth_cache_path(), client$name)
   dir.create(app_path, showWarnings = FALSE, recursive = TRUE)
 
   path <- file.path(app_path, paste0(hash(key), "-token.rds.enc"))
@@ -109,7 +109,7 @@ cache_disk <- function(client, key) {
 }
 
 # Update req_oauth_auth_code() docs if change default from 30
-cache_disk_prune <- function(days = 30, path = rappdirs::user_cache_dir("httr2")) {
+cache_disk_prune <- function(days = 30, path = oauth_cache_path()) {
   files <- dir(path, recursive = TRUE, full.names = TRUE, pattern = "-token\\.rds$")
   mtime <- file.mtime(files)
 
@@ -117,16 +117,15 @@ cache_disk_prune <- function(days = 30, path = rappdirs::user_cache_dir("httr2")
   unlink(files[old])
 }
 
-#' httr2 cache location
+#' httr2 OAuth cache location
 #'
 #' When opted-in to, httr2 caches OAuth tokens in this directory. By default,
 #' it uses a OS-standard cache directory, but, if needed, you can override the
-#' location by setting the `HTTR2_CACHE` env var.
+#' location by setting the `HTTR2_OAUTH_CACHE` env var.
 #'
 #' @export
-#' @keywords internal
-cache_path <- function() {
-  path <- Sys.getenv("HTTR2_CACHE")
+oauth_cache_path <- function() {
+  path <- Sys.getenv("HTTR2_OAUTH_CACHE")
   if (path != "") {
     return(path)
   }